Tests/LibWeb/Text/expected/navigate-attachment-downloads.txt

attachment requested a download: true
